refact: move util functions to own file
